Added simple auth. Added role table and role_id row to user. Added policy to quotes.

+++ app/Http/Controllers/QuoteController.php
@@ -8,6 +8,7 @@ use App\Http\Requests\StoreQuoteRequest;
 use App\Http\Resources\QuoteResource;
 use App\Models\Quote;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class QuoteController extends Controller
 {
@@ -16,14 +17,14 @@ class QuoteController extends Controller
      */
     public function index(Request $request)
     {
-        $query = Quote::with('tags');
+        $query = Quote::with('tags')->orderBy('created_at', 'desc');
 
         if ($request->has('search')) {
             $search = $request->search;
 
             $query->where(function ($q) use ($search) {
-               $q->where('title', 'like', '%' . $search . '%')
-                   ->orWhere('text', 'like', '%' . $search . '%');
+                $q->where('title', 'like', '%'.$search.'%')
+                    ->orWhere('text', 'like', '%'.$search.'%');
             });
         }
         $quotes = $query->paginate(10);
@@ -73,7 +74,7 @@ class QuoteController extends Controller
             $quote = UpdateQuoteAction::execute($request->validated(), $quote);
         } catch (\Exception $e) {
             return response()->json([
-            'message' => $e->getMessage(),
+                'message' => $e->getMessage(),
             ], 422);
         }
 
@@ -83,18 +84,20 @@ class QuoteController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $id)
+    public function destroy(Request $request, string $id)
     {
         try {
             $quote = Quote::findOrFail($id);
+            Gate::authorize('delete', $quote);
+
             $quote->delete();
 
             return response()->json([
-                'message' => 'Цитата успешно удалена.'
+                'message' => 'Цитата успешно удалена.',
             ]);
         } catch (\Exception $e) {
             return response()->json([
-                'message' => 'Ошибка, такая цитата не существует.',
+                'message' => $e->getMessage(),
             ], 404);
         }
     }
